+++ backend/tests/Functional/Controller/App/ApplicationControllerTest.php
@@ -40,7 +40,7 @@ class ApplicationControllerTest extends WebTestCase
             'id' => $app->getId(),
             'name' => 'Test App',
             'groups' => [['id' => $group->getId(), 'name' => 'g1', 'description' => null,
-                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false]],
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false]],
             'roles' => ['app'],
             'eveLogins' => [],
         ], $this->parseJsonBody($response));